+++ src/blocks/World.hpp
@@ -15,7 +15,7 @@ class World
 	// Define hasher for glm::vec2 used for unordered_map (activeChunks)
 	struct Vec2Hasher
 	{
-		std::size_t operator()(const glm::vec2& c) const
+		std::size_t operator()(const glm::ivec2& c) const
 		{
 			// XOR hashes of x and z in chunkPos
 			return std::hash<int>()(c.x) ^ (std::hash<int>()(c.y) << 1);
@@ -24,7 +24,7 @@ class World
 
 	struct Vec2Equals
 	{
-		bool operator()(const glm::vec2& lhs, const glm::vec2& rhs) const
+		bool operator()(const glm::ivec2& lhs, const glm::ivec2& rhs) const
 		{
 			return lhs.x == rhs.x && lhs.y == rhs.y;
 		}
@@ -34,8 +34,8 @@ private:
 
 	// Chunk data
 	std::unique_ptr<Chunk> spawnChunk;
-	std::unordered_map<glm::vec2, std::unique_ptr<Chunk>, Vec2Hasher, Vec2Equals> activeChunks;
-	std::unordered_set<glm::vec2, Vec2Hasher, Vec2Equals> dirtyChunks;
+	std::unordered_map<glm::ivec2, std::unique_ptr<Chunk>, Vec2Hasher, Vec2Equals> activeChunks;
+	std::unordered_set<glm::ivec2, Vec2Hasher, Vec2Equals> dirtyChunks;
 	
 	FastNoiseLite perlinNoise;
